Add the npnreload command
The `npnreload`-command is used for the reload of the config.yml so that the bungee does not need to restart.

The command itself does not get autocompleted when a player does not have the permission for the command: "noplayernotifier.reload".

The command can only be executed with op on the bungee (console) or when the player has the permission.